remove GridOverlays export
