+++ packages/x-data-grid/src/components/base/index.ts
@@ -1,3 +1,2 @@
 export * from './GridBody';
 export * from './GridFooterPlaceholder';
-export * from './GridOverlays';
